BandEntity Corrigido
