+++ app/src/main/java/com/rogacheski/lbd/lbdmusic/MainActivity.java
@@ -1,17 +1,11 @@
 package com.rogacheski.lbd.lbdmusic;
 
 import android.app.ActivityManager;
-import android.content.Context;
-import android.content.DialogInterface;
 import android.content.Intent;
 import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
 import android.graphics.Color;
-import android.graphics.PorterDuff;
-import android.graphics.drawable.Drawable;
 import android.os.Bundle;
-import android.support.design.widget.FloatingActionButton;
-import android.support.design.widget.Snackbar;
 import android.util.Log;
 import android.view.KeyEvent;
 import android.view.View;
@@ -19,7 +13,6 @@ import android.support.design.widget.NavigationView;
 import android.support.v4.view.GravityCompat;
 import android.support.v4.widget.DrawerLayout;
 import android.support.v7.app.ActionBarDrawerToggle;
-import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.Toolbar;
 import android.view.Menu;
 import android.view.MenuItem;
@@ -30,7 +23,6 @@ import com.rogacheski.lbd.lbdmusic.base.baseActivity;
 import com.rogacheski.lbd.lbdmusic.entity.BandEntity;
 import com.rogacheski.lbd.lbdmusic.session.Session;
 import com.rogacheski.lbd.lbdmusic.model.user;
-import com.rogacheski.lbd.lbdmusic.model.band;
 import com.rogacheski.lbd.lbdmusic.singleton.PicassoSingleton;
 
 import android.view.inputmethod.EditorInfo;
@@ -45,7 +37,6 @@ import org.json.JSONObject;
 
 import java.lang.ref.WeakReference;
 
-import butterknife.Bind;
 import butterknife.ButterKnife;
 import cz.msebera.android.httpclient.Header;
