+++ app/src/main/java/com/rogacheski/lbd/lbdmusic/searchResult.java
@@ -7,12 +7,8 @@ import android.os.Bundle;
 import android.view.View;
 import android.widget.AdapterView;
 import android.widget.ArrayAdapter;
-import android.widget.EditText;
 import android.widget.ListView;
-import android.widget.RelativeLayout;
 
-import com.rogacheski.lbd.lbdmusic.entity.BandEntity;
-import com.rogacheski.lbd.lbdmusic.model.band;
 import com.rogacheski.lbd.lbdmusic.model.user;
 
 import java.util.ArrayList;
